Split reactions and molecules in the chemdraw support libs
